requirements.txt
openpyxl==3.1.2
python-dotenv==1.0.0
WooCommerce==3.0.0
xlrd==2.0.1
